+++ axum-extra/src/extract/mod.rs
@@ -1,10 +1,29 @@
 //! Additional extractors.
 
 mod cached;
+
+#[cfg(feature = "form")]
+mod form;
+
 #[cfg(feature = "cookie")]
 pub mod cookie;
+
+#[cfg(feature = "query")]
+mod query;
 
 pub use self::cached::Cached;
 
 #[cfg(feature = "cookie")]
-pub use self::cookie::{CookieJar, PrivateCookieJar, SignedCookieJar};
+pub use self::cookie::CookieJar;
+
+#[cfg(feature = "cookie-private")]
+pub use self::cookie::PrivateCookieJar;
+
+#[cfg(feature = "cookie-signed")]
+pub use self::cookie::SignedCookieJar;
+
+#[cfg(feature = "form")]
+pub use self::form::Form;
+
+#[cfg(feature = "query")]
+pub use self::query::Query;