+++ axum-macros/tests/debug_handler/fail/multiple_body_extractors.rs
@@ -1,0 +1,7 @@
+use axum_macros::debug_handler;
+use axum::body::Bytes;
+
+#[debug_handler]
+async fn handler(_: String, _: Bytes) {}
+
+fn main() {}