+++ axum/src/extract/form.rs
@@ -58,7 +58,7 @@ where
         if req.method() == Method::GET {
             let query = req.uri().query().unwrap_or_default();
             let value = serde_urlencoded::from_str(query)
-                .map_err(FailedToDeserializeQueryString::new::<T, _>)?;
+                .map_err(FailedToDeserializeQueryString::__private_new::<T, _>)?;
             Ok(Form(value))
         } else {
             if !has_content_type(req, &mime::APPLICATION_WWW_FORM_URLENCODED) {
@@ -67,7 +67,7 @@ where
 
             let bytes = Bytes::from_request(req).await?;
             let value = serde_urlencoded::from_bytes(&bytes)
-                .map_err(FailedToDeserializeQueryString::new::<T, _>)?;
+                .map_err(FailedToDeserializeQueryString::__private_new::<T, _>)?;
 
             Ok(Form(value))
         }