+++ axum/src/extract/rejection.rs
@@ -64,6 +64,14 @@ define_rejection! {
 }
 
 define_rejection! {
+    #[status = BAD_REQUEST]
+    #[body = "`GET`, `HEAD`, `OPTIONS` requests are not allowed to have a `Content-Length` header"]
+    /// Rejection type for [`ContentLengthLimit`](super::ContentLengthLimit) if
+    /// the request is `GET`, `HEAD`, or `OPTIONS` and has a `Content-Length` header.
+    pub struct ContentLengthNotAllowed;
+}
+
+define_rejection! {
     #[status = INTERNAL_SERVER_ERROR]
     #[body = "No paths parameters found for matched route. Are you also extracting `Request<_>`?"]
     /// Rejection type used if axum's internal representation of path parameters
@@ -74,7 +82,7 @@ define_rejection! {
 
 define_rejection! {
     #[status = UNSUPPORTED_MEDIA_TYPE]
-    #[body = "Form requests must have `Content-Type: x-www-form-urlencoded`"]
+    #[body = "Form requests must have `Content-Type: application/x-www-form-urlencoded`"]
     /// Rejection type used if you try and extract the request more than once.
     pub struct InvalidFormContentType;
 }
@@ -96,7 +104,8 @@ pub struct FailedToDeserializeQueryString {
 }
 
 impl FailedToDeserializeQueryString {
-    pub(super) fn new<T, E>(error: E) -> Self
+    #[doc(hidden)]
+    pub fn __private_new<T, E>(error: E) -> Self
     where
         E: Into<BoxError>,
     {
@@ -225,6 +234,8 @@ pub enum ContentLengthLimitRejection<T> {
     #[allow(missing_docs)]
     LengthRequired(LengthRequired),
     #[allow(missing_docs)]
+    ContentLengthNotAllowed(ContentLengthNotAllowed),
+    #[allow(missing_docs)]
     Inner(T),
 }
 
@@ -236,6 +247,7 @@ where
         match self {
             Self::PayloadTooLarge(inner) => inner.into_response(),
             Self::LengthRequired(inner) => inner.into_response(),
+            Self::ContentLengthNotAllowed(inner) => inner.into_response(),
             Self::Inner(inner) => inner.into_response(),
         }
     }
@@ -249,6 +261,7 @@ where
         match self {
             Self::PayloadTooLarge(inner) => inner.fmt(f),
             Self::LengthRequired(inner) => inner.fmt(f),
+            Self::ContentLengthNotAllowed(inner) => inner.fmt(f),
             Self::Inner(inner) => inner.fmt(f),
         }
     }
@@ -262,6 +275,7 @@ where
         match self {
             Self::PayloadTooLarge(inner) => Some(inner),
             Self::LengthRequired(inner) => Some(inner),
+            Self::ContentLengthNotAllowed(inner) => Some(inner),
             Self::Inner(inner) => Some(inner),
         }
     }