+++ axum/src/extract/ws.rs
@@ -207,12 +207,17 @@ impl WebSocketUpgrade {
         let on_upgrade = self.on_upgrade;
         let config = self.config;
 
+        let protocol = self.protocol.clone();
+
         tokio::spawn(async move {
             let upgraded = on_upgrade.await.expect("connection upgrade failed");
             let socket =
                 WebSocketStream::from_raw_socket(upgraded, protocol::Role::Server, Some(config))
                     .await;
-            let socket = WebSocket { inner: socket };
+            let socket = WebSocket {
+                inner: socket,
+                protocol,
+            };
             callback(socket).await;
         });
 
@@ -309,6 +314,7 @@ fn header_contains<B>(req: &RequestParts<B>, key: HeaderName, value: &'static st
 #[derive(Debug)]
 pub struct WebSocket {
     inner: WebSocketStream<Upgraded>,
+    protocol: Option<HeaderValue>,
 }
 
 impl WebSocket {
@@ -330,6 +336,11 @@ impl WebSocket {
     /// Gracefully close this WebSocket.
     pub async fn close(mut self) -> Result<(), Error> {
         self.inner.close(None).await.map_err(Error::new)
+    }
+
+    /// Return the selected WebSocket subprotocol, if one has been chosen.
+    pub fn protocol(&self) -> Option<&HeaderValue> {
+        self.protocol.as_ref()
     }
 }